+++ app/services/payments/providers/base.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
+from decimal import Decimal
 
 
 @dataclass
@@ -15,10 +16,10 @@ class ProviderPaymentResult:
 class PaymentProvider:
     name: str = "base"
 
-    async def create_intent(self, *, amount: float, currency: str) -> ProviderPaymentResult:
+    async def create_intent(self, *, amount: Decimal, currency: str) -> ProviderPaymentResult:
         raise NotImplementedError
 
     async def capture(
-        self, *, external_id: str, amount: float, idempotency_key: str
+        self, *, external_id: str, amount: Decimal, idempotency_key: str
     ) -> ProviderPaymentResult:
         raise NotImplementedError
